refactor: moved helpers
